src/server/requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
flake8==3.9.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.0.1
PyMySQL==1.0.2
python-dotenv==0.14.0
regex==2021.3.17
SQLAlchemy==1.4.5
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
